Move tests to a new file per PR feedback
